remove version check and test
