+++ internal/k8s/deployments.go
@@ -17,6 +17,8 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -289,4 +291,24 @@ func NewDeploymentWithOwnerRef(
 			MinReadySeconds: spec.MinReadySeconds,
 		},
 	}
+}
+
+func NewDeploymentWithControllerRef(
+	w *temporaliov1alpha1.TemporalWorkerDeployment,
+	buildID string,
+	connection temporaliov1alpha1.TemporalConnectionSpec,
+	reconcilerScheme *runtime.Scheme,
+) (*appsv1.Deployment, error) {
+	d := NewDeploymentWithOwnerRef(
+		&w.TypeMeta,
+		&w.ObjectMeta,
+		&w.Spec,
+		ComputeWorkerDeploymentName(w),
+		buildID,
+		connection,
+	)
+	if err := ctrl.SetControllerReference(w, d, reconcilerScheme); err != nil {
+		return nil, err
+	}
+	return d, nil
 }
